changed product back to use BigDecimal

+++ jmaqs-webservices-jdk8/src/test/java/com/magenic/jmaqs/webservices/jdk8/models/Product.java
@@ -6,6 +6,7 @@ package com.magenic.jmaqs.webservices.jdk8.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+import java.math.BigDecimal;
 
 public class Product {
     /**
@@ -24,9 +25,9 @@ public class Product {
     private String category;
 
     @JsonProperty("Price")
-    private double price;
+    private BigDecimal price;
 
-    public Product(int id, String name, String category, double price) {
+    public Product(int id, String name, String category, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.category = category;
@@ -59,11 +60,11 @@ public class Product {
         this.category = category;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal  price) {
         this.price = price;
     }
